+++ Src/Graphics/GraphicsResources.h
@@ -25,7 +25,7 @@ class GraphicsResources {
             bool needsViewProjection;
         };
         std::map<PGE::Shader*, ShaderEntry*> shaderToShaders;
-        std::map<long long, ShaderEntry*> pathToShaders;
+        std::map<uint64_t, ShaderEntry*> pathToShaders;
 
         struct TextureEntry {
             // This needs to stay a string for the Resource Packs to work.
@@ -34,7 +34,7 @@ class GraphicsResources {
             int refCount;
         };
         std::map<PGE::Texture*, TextureEntry*> textureToTextures;
-        std::map<long long, TextureEntry*> pathToTextures;
+        std::map<uint64_t, TextureEntry*> pathToTextures;
 
         struct ModelEntry {
             // Having this as a string makes the loading of textures easier.
@@ -43,7 +43,7 @@ class GraphicsResources {
             int refCount;
         };
         std::map<Model*, ModelEntry*> modelToModels;
-        std::map<long long, ModelEntry*> pathToModels;
+        std::map<uint64_t, ModelEntry*> pathToModels;
 
         PGE::Matrix4x4f orthoMat;
